Routes: Show Page
The following steps have been completed for all applications.

Add a generic route to handle requests for /movies/1, /movies/2,
/movies/3, and so on.

Define a show action in the MoviesController that finds the movie with
the id specified in the URL.

Create a show.html.erb template that generates HTML to display the
movie's details.